Move to Shared for SqlConnectionPoolKey.cs (#1292)
